fix: keep Span at 16 bytes to avoid parser stack overflows
Adding end positions doubled Span to 4x usize (32 bytes). Span is
embedded in every token and AST node, so the recursive-descent parser's
stack frames grew enough that deep-but-legitimate recursion (builtin
prelude expansion) overflowed the 2 MiB test-thread stacks, aborting
hundreds of tests with SIGABRT. Coordinates are now u32 (like rustc's
spans), restoring the original 16-byte size; a size regression test
pins it.

+++ src/lexer/cursor.rs
@@ -50,7 +50,7 @@ impl<'a> Cursor<'a> {
     ///
     /// Used to attach source locations to tokens for diagnostics.
     pub fn span(&self) -> Span {
-        Span::new(self.line, self.col)
+        Span::new(self.line as u32, self.col as u32)
     }
 
     /// Returns the next character without advancing the cursor.
